feat: add update handler and delete check for 404 when id does not valid

+++ src/controller/todo.rs
@@ -5,11 +5,12 @@ use axum::{
     extract::{Path, State},
     http::StatusCode,
 };
+use chrono::Utc;
 use serde_json::{Value, json};
 
 use crate::{
     app_state::AppState,
-    models::todo::{CreateTodo, Todo},
+    models::todo::{CreateTodo, Todo, UpdateTodo},
 };
 
 // region :      --- Todo Handler
@@ -20,15 +21,15 @@ pub async fn list_todo(
         .fetch_all(&state.connection)
         .await;
     match result {
-        Ok(todo) => Ok(Json(json!({ "data": todo }))),
+        Ok(todo) => return Ok(Json(json!({ "data": todo }))),
         Err(e) => {
             tracing::error!("Database error: {:?}", e);
-            Err((
+            return Err((
                 StatusCode::INTERNAL_SERVER_ERROR,
                 Json(json!({
                     "error": "internal server error"
                 })),
-            ))
+            ));
         }
     }
 }
@@ -44,15 +45,68 @@ pub async fn add_todo(
             .await;
 
     match result {
-        Ok(id) => Ok((StatusCode::CREATED, Json(json!({ "id": id })))),
+        Ok(id) => return Ok((StatusCode::CREATED, Json(json!({ "id": id })))),
         Err(e) => {
             tracing::error!("Database error: {:?}", e);
-            Err((
+            return Err((
                 StatusCode::INTERNAL_SERVER_ERROR,
                 Json(json!({
                     "error": "internal server error"
                 })),
-            ))
+            ));
+        }
+    }
+}
+
+pub async fn update_todo(
+    State(state): State<Arc<AppState>>,
+    Path(id): Path<i64>,
+    Json(payload): Json<UpdateTodo>,
+) -> Result<(StatusCode, Json<Value>), (StatusCode, Json<Value>)> {
+    // TODO: refactor this update_todo and delete_todo both are using this statement.
+    let record = sqlx::query_as::<_, Todo>(r#"select * from todos where id=($1)"#)
+        .bind(id)
+        .fetch_optional(&state.connection)
+        .await
+        .map_err(|e| {
+            tracing::error!("Database error: {:?}", e);
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({
+                    "error": "internal server error"
+                })),
+            );
+        })?;
+    // TODO: refactor this update_todo and delete_todo both are using this statement.
+    if record.is_none() {
+        tracing::error!("Todo ID:{} does not exists", id);
+        return Err((
+            StatusCode::NOT_FOUND,
+            Json(json!({
+                "error": "record does not exists"
+            })),
+        ));
+    }
+    // UPDATE ROW VIA SQL
+    let updated_at = Utc::now();
+    let result = sqlx::query_scalar::<_, i64>(r#"update todos set description=($1), completed=($2), updated_at=($3) where id=($4) returning id"#)
+    .bind(payload.description)
+    .bind(payload.completed)
+    .bind(updated_at)
+    .bind(id)
+    .fetch_one(&state.connection)
+    .await;
+    // Return Result
+    match result {
+        Ok(id) => return Ok((StatusCode::OK, Json(json!({ "id": id })))),
+        Err(e) => {
+            tracing::error!("Database error: {:?}", e);
+            return Err((
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({
+                    "error": "internal server error"
+                })),
+            ));
         }
     }
 }
@@ -61,21 +115,46 @@ pub async fn delete_todo(
     State(state): State<Arc<AppState>>,
     Path(id): Path<i64>,
 ) -> Result<Json<Value>, (StatusCode, Json<Value>)> {
-    let result = sqlx::query(r#"delete from todos where id=($1);"#)
+    // TODO: refactor this
+    let record = sqlx::query_as::<_, Todo>(r#"select * from todos where id=($1)"#)
+        .bind(id)
+        .fetch_optional(&state.connection)
+        .await
+        .map_err(|e| {
+            tracing::error!("Database error: {:?}", e);
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({
+                    "error": "internal server error"
+                })),
+            );
+        })?;
+    // TODO: refactor this
+    if record.is_none() {
+        tracing::error!("Todo ID:{} does not exists", id);
+        return Err((
+            StatusCode::NOT_FOUND,
+            Json(json!({
+                "error": "record does not exists"
+            })),
+        ));
+    }
+
+    let result = sqlx::query(r#"delete from todos where id=($1)"#)
         .bind(id)
         .execute(&state.connection)
         .await;
 
     match result {
-        Ok(_) => Ok(Json(json!({ "message": "success" }))),
+        Ok(_) => return Ok(Json(json!({ "message": "success" }))),
         Err(e) => {
             tracing::error!("Database error: {:?}", e);
-            Err((
+            return Err((
                 StatusCode::INTERNAL_SERVER_ERROR,
                 Json(json!({
                     "error": "internal server error"
                 })),
-            ))
+            ));
         }
     }
 }
